# bci-hackathon/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/register/":{"post":{"summary":"Register Device","operationId":"register_device_register__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Device"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConnectionDetails"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/last/session/{minutes}":{"get":{"summary":"Getlastsessiondata","operationId":"getLastSessionData_last_session__minutes__get","parameters":[{"name":"minutes","in":"path","required":true,"schema":{"type":"integer","title":"Minutes"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SessionData"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ConnectionDetails":{"properties":{"connected":{"type":"boolean","title":"Connected"},"battery":{"type":"integer","title":"Battery"}},"type":"object","required":["connected","battery"],"title":"ConnectionDetails"},"Device":{"properties":{"serialNumber":{"type":"string","title":"Serialnumber"},"model":{"type":"string","title":"Model"}},"type":"object","required":["serialNumber","model"],"title":"Device"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Raw":{"properties":{"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"f4":{"type":"number","title":"F4"},"f3":{"type":"number","title":"F3"},"c4":{"type":"number","title":"C4"},"c3":{"type":"number","title":"C3"},"p4":{"type":"number","title":"P4"},"p3":{"type":"number","title":"P3"},"o1":{"type":"number","title":"O1"},"o2":{"type":"number","title":"O2"}},"type":"object","required":["timestamp","f4","f3","c4","c3","p4","p3","o1","o2"],"title":"Raw"},"Scores":{"properties":{"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"coef_min":{"type":"number","title":"Coef Min"},"coef_max":{"type":"number","title":"Coef Max"},"coef_avg":{"type":"number","title":"Coef Avg"},"is_focused":{"type":"boolean","title":"Is Focused"}},"type":"object","required":["timestamp","coef_min","coef_max","coef_avg","is_focused"],"title":"Scores"},"SessionData":{"properties":{"scores":{"items":{"$ref":"#/components/schemas/Scores"},"type":"array","title":"Scores"},"raw":{"items":{"$ref":"#/components/schemas/Raw"},"type":"array","title":"Raw"}},"type":"object","required":["scores","raw"],"title":"SessionData"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}